Added Product

+++ src/main/java/com/vodafone/Online_Store/infrastructure/repository/UserInfoRepository.java
@@ -1,11 +1,14 @@
 package com.vodafone.Online_Store.infrastructure.repository;
 
+import com.vodafone.Online_Store.core.domain.UserInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.vodafone.Online_Store.core.domain.UserInfo;
+
 import java.util.Optional;
 
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
+    // Additional custom queries if needed
     Optional<UserInfo> findByUsername(String username);
+
 }
